refactor: Far too large a commit, but all intertwined (see below)
Commit is mostly getting a handle on flow types and organizing the types
into their own files in a types directory. The type files correspond to
new action and reducer files for each slice of the state. Component and
container files changed to import new files instead of old ones.