refactor: Restructure how app loads, displays, sorts, and filters posts
App is now a container component responsible for fetching all posts on
mounting and then rendering the main list of posts route and the single
post route. ListOfPosts container handles category filtering, post
sorting, and rendering the PostList, SelectCategory, and PostSort
components. ListOfPosts handles category filtering by getting the
cagtegory parameter from the url.

The asynchronous action creator called within App's ComponentDidMount
dispatches a GET_ALL_POSTS_STARTED and a GET_ALL_POSTS_SUCCEEDED on
resolution of promise in order to allow for a postsLoading state to be
tracked. UI has a loading indicator that displays to indicate that posts
are loading to user.