fix: Change timing of loader displays to avoid accessing posts before they load
There is alot of rearranging in commit, but the key part is in
ui-reducer.js where postsLoading is set to true by default. Prior to
calling getAllPosts, which would result in postsLoading being set to
true, there was a brief period where postsLoading was false. This meant
that the loaders meant to delay trying to render posts until they loaded
did not display. Without the loaders, the components would render and
cause type errors for trying to access posts that hadn't loaded yet. By
setting the postsLoading to true by default, the loaders display in the
initial page render and the type errors don't occur.