fix: Refactor how comments are loaded and state is tracked
Initially comments were fetched from server by PostCardDisplay and then
local state was set in ComponentDidMount to track and caculate number of
comments each post had. This resulted in console warning for calling
setState on unmounted components. To fix, stopped using local state and
instead changed the structure of how comments in the store to
facillitate easier calculation of the number of comments per post.